seperated out the styles and segregrated the code into sub components
